refactoring the libs
