+++ incl/fixed_types.h
@@ -1,6 +1,5 @@
 #ifndef FIXED_TYPES_H
 # define FIXED_TYPES_H
-# include "cub.h"
 
 typedef unsigned int t_fixed8;
