+++ incl/map_parsing.h
@@ -1,4 +1,5 @@
 #ifndef MAP_PARSING_H
 # define MAP_PARSING_H
+# include "cub.h"
 
 #endif //CUB3D_IN_LOVE_MAP_PARSING_H
