+++ srcs/libs/libft/colors.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/cub.h"
 
 void	white()
 {
