+++ srcs/libs/libft/freematrix.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/cub.h"
 
 void	freematrix(char **splitted, int i)
 {
