+++ srcs/libs/libft/ft_isascii.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 int	ft_isascii(int c)
 {
