+++ srcs/libs/libft/ft_isnum.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "../../includes/libft.h"
+#include "../../../incl/libft.h"
 
 int	ft_isnum(char *str)
 {
