+++ srcs/libs/libft/ft_isprint.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 int	ft_isprint(int c)
 {
