+++ srcs/libs/libft/ft_isspace.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_isspace(int c) 
+#include "../../../incl/libft.h"
+
+int	ft_isspace(int c)
 {
 	if (c == ' ' || (c > 8 && c < 14))
 		return (1);
-	return (0);	
+	return (0);
 }
