+++ srcs/libs/libft/ft_split.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 static int	wc(const char *s, char c)
 {
