+++ srcs/libs/libft/ft_strchr.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
