+++ srcs/libs/libft/ft_strcmp.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 int	ft_strcmp(char *s1, char *s2)
 {
