+++ srcs/libs/libft/ft_strjoin.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 char	*ft_strjoin(const char *s1, const char *s2)
 {
