+++ srcs/libs/libft/ft_substr.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/libft.h"
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
