+++ srcs/libs/listft/del_node.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/listft.h"
 
 static void	del_node_null(t_node *current, t_node **stack)
 {
