+++ srcs/libs/listft/list_point_to.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/listft.h"
 
 void	list_point_to(t_node **list, int id)
 {
