+++ srcs/libs/listft/matrix_to_list.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/listft.h"
 
 t_node	*matrix_to_list(char **matrix)
 {
