+++ srcs/libs/listft/new_.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/program.h"
+#include "../../../incl/listft.h"
 
 t_node	*new_(char *content, int id)
 {
