+++ srcs/libs/listft/src_in_list.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/libft.h"
+#include "../../../incl/listft.h"
 
 int	src_in_list(t_node	**h, char *field)
 {
